requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-dev --no-emit-project -o requirements.txt
anyio==4.12.1
    # via starlette
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.12.0
    # via functions-framework
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==46.0.5
    # via google-auth
deprecation==2.1.0
    # via cloudevents
flask==3.1.2
    # via
    #   functions-framework
    #   gfc-redirect
functions-framework==3.10.0
    # via gfc-redirect
google-auth==2.48.0
    # via
    #   gfc-redirect
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.4
    # via gspread
gspread==6.2.1
    # via gfc-redirect
gunicorn==25.1.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via
    #   deprecation
    #   gunicorn
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
requests==2.32.5
    # via requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
starlette==0.52.1
    # via functions-framework
typing-extensions==4.15.0 ; python_full_version < '3.13'
    # via
    #   anyio
    #   starlette
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.5
    # via
    #   flask
    #   functions-framework
